[FIX] 폴더구조 고치기
