+++ gayoung/notes.md
