+++ gayoung/week1/codes/my-app/src/App.css
@@ -2,10 +2,6 @@
   text-align: center;
 }
 
-p {
-  color: skyblue;
-}
-
 .App-logo {
   height: 40vmin;
   pointer-events: none;
@@ -25,7 +21,7 @@ p {
   align-items: center;
   justify-content: center;
   font-size: calc(10px + 2vmin);
-  color: white;
+  color: #61dafb;
 }
 
 .App-link {
