+++ gayoung/week1/codes/my-app/src/App.js
@@ -1,0 +1,15 @@
+import logo from "./logo.svg";
+import "./App.css";
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>리액트를 시작해봐요!</p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
